Outline Shaders + new textures for placeholder goblin model
There's a new goblin model present in the game. It's somewhat ready for IK stuff, the rig hasn't changed -- but the head is detached from the rest of the mesh, this was for some speculative development. I'm unsure how it works but just know that it's there and potentially will change.